+++ core/src/main/java/com/hujiugame/qingfeng/type/key/VersionKey.java
@@ -14,11 +14,11 @@ public final class VersionKey
     // app_version.json 字段
 
     /** 内部版本号 */
-    public static final String APP_VERSION = "appVersion";
+    public static final String APP_VERSION = "app_version";
     /** 内部版本类型 */
-    public static final String APP_VERSION_TYPE = "appVersionType";
+    public static final String APP_VERSION_TYPE = "app_version_type";
     /** 内部版本字符串 */
-    public static final String APP_VERSION_STRING = "appVersionString";
+    public static final String APP_VERSION_STRING = "app_version_string";
 
     // ====================================================================================================
     // 远程 versions.json 字段